requirements.txt
pandas~=1.4.2
PyPDF2~=1.27.5
pytesseract~=0.3.9
bs4~=0.0.1
requests~=2.27.1
pdfplumber~=0.6.0
streamlit~=1.8.1
docx2txt~=0.8
beautifulsoup4~=4.11.1
soupsieve~=2.3.2.post1
numpy~=1.22.3
pip~=21.3.1
pyzmq~=22.3.0
six~=1.16.0
attrs~=21.4.0
click~=8.0.4
toml~=0.10.2
colorama~=0.4.4
gitdb~=4.0.9
parso~=0.8.3
toolz~=0.11.2
altair~=4.2.0
ipython~=8.2.0
jsonschema~=4.4.0
Jinja2~=3.1.1
entrypoints~=0.4
webencodings~=0.5.1
packaging~=21.3
bleach~=5.0.0
Pillow~=9.1.0
tornado~=6.1
validators~=0.18.2
blinker~=1.4
certifi~=2021.10.8
chardet~=4.0.0
debugpy~=1.6.0
traitlets~=5.1.1
setuptools~=60.2.0
decorator~=5.1.1
backcall~=0.2.0
python-dateutil~=2.8.2
notebook~=6.4.10
Pygments~=2.11.2
asttokens~=2.0.5
executing~=0.8.3
nbconvert~=6.5.0
pyrsistent~=0.18.1
wcwidth~=0.2.5
cryptography~=36.0.2
pyparsing~=3.0.8
psutil~=5.9.0
trafilatura~=1.2.0